Added Locatable Interface and Refactored GeoPoint Class to an Interface
I added a Locatable interface to the com.wildstartech.wfa.location
package which provides the ability to retrieve a Location object from a
class.  The Location object is the parent object for all location
objects.  In this initial phase, there are NO properties for the
Location object.  It is essentially a marker interface.

I also took the GeoPoint class and converted it into an interface.  The
GeoPointImpl class was created in the
com.wildstartech.wfa.location.address.us package.

As the GeoPoint class featured a calculateDistance static method which
could not be present in an interface, the DistanceUtilities class was
created and the static calculateDistance method was moved to that class.

Also created a Warehouse location which is a sub-class of the Location
object and exists in the com.wildstartech.wfa.logistics package.